Update profile editing
* Move ProfilePage to ProfileEditorPage
* Add preferred rating system, fide id and uscf id

Related to #45
