+++ backend/database/user.go
@@ -65,6 +65,9 @@ type User struct {
 	// The user's Discord username
 	DiscordUsername string `dynamodbav:"discordUsername,omitempty" json:"discordUsername,omitempty"`
 
+	// The user's preferred rating system
+	RatingSystem string `dynamodbav:"ratingSystem" json:"ratingSystem"`
+
 	// The user's Chess.com username
 	ChesscomUsername string `dynamodbav:"chesscomUsername,omitempty" json:"chesscomUsername,omitempty"`
