+++ frontend/src/App.tsx
@@ -5,7 +5,7 @@ import { ThemeProvider } from '@mui/system';
 import { getConfig } from './config';
 import { AuthProvider, RequireAuth, RequireProfile } from './auth/Auth';
 import LandingPage from './home/LandingPage';
-import ProfilePage from './profile/ProfilePage';
+import ProfileEditorPage from './profile/ProfileEditorPage';
 import { ApiProvider } from './api/Api';
 import CalendarPage from './calendar/CalendarPage';
 import MeetingPage from './meeting/MeetingPage';
@@ -69,7 +69,7 @@ function Router() {
                 <Route path='forgot-password' element={<ForgotPasswordPage />} />
 
                 <Route element={<RequireAuth />}>
-                    <Route path='profile' element={<ProfilePage />} />
+                    <Route path='profile/edit' element={<ProfileEditorPage />} />
                     <Route path='admin' element={<AdminPage />} />
 
                     <Route element={<RequireProfile />}>
